Create model + questions manager
